+++ src/server/settings.tsx
@@ -1,2 +1,2 @@
-import settings from '../../config/server';
-export default settings;
+import settings from "../config/store"
+export default settings
