+++ tsconfig.json
@@ -9,7 +9,8 @@
     "target": "es2017",
     "esModuleInterop": true,
     "noEmitOnError": false,
-    "removeComments": true
+    "removeComments": true,
+    "jsx": "react"
   },
   "compileOnSave": true,
   "include": ["src"]
